requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
pydantic
python-dotenv==1.0.1
httpx==0.28.1
requests 
livekit-agents[openai, silero, elevenlabs,groq]==1.0.11
python-dotenv==1.0.1
httpx==0.28.1
ipython==8.13.2
google-generativeai
livekit-agents[google]~=1.0
livekit-plugins-noise-cancellation~=0.2
serpapi
aiohttp
sqlalchemy